refactor: use SynonFetchContext in Synonym practice - seperate into SearchInput & SynonList component
